+++ starter-code/src/components/App/App.css
@@ -1,0 +1,17 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+body {
+  background-color: rgb(160, 139, 102);
+  height: 100vh;
+}
+
+.mainBox {
+  display: flex;
+  justify-content: space-around;
+  align-content: center;
+  margin: 40px;
+}
