+++ starter-code/src/components/Container/Container.css
@@ -1,0 +1,5 @@
+.message {
+  width: 30vw;
+  height: fit-content;
+  box-shadow: 2px 2px 15px  rgba(0, 0, 0, 0.192);
+}
