+++ starter-code/src/components/Container/Container.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
-import Message from './Message'
+import Message from '../Message/Message'
+import './Container.css'
 
 export default class Container extends Component {
   constructor() {
@@ -8,7 +9,7 @@ export default class Container extends Component {
 
   render() {
     return (
-      <div>
+      <div class='message'>
         <Message isInfo title="Hello World">Lorem ipsum dolor sit amet, consectetur adipiscing elit. <strong>Pellentesque risus mi</strong>.</Message>
       </div>
     )
