+++ starter-code/src/components/NavBar/NavBar.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
-import CoolButton from "./CoolButton"
+import CoolButton from "../CoolButton/CoolButton"
+import './NavBar.css'
 
 export default class NavBar extends Component {
   constructor() {
