+++ starter-code/src/components/SignUp/SignUp.css
@@ -1,0 +1,6 @@
+.SignUp {
+  width: 30vw;
+  background-color: rgba(255, 255, 255, 0.20);
+  padding: 20px 40px;
+  box-shadow: 2px 2px 5px  rgba(0, 0, 0, 0.493);
+}
