+++ starter-code/src/index.js
@@ -1,7 +1,7 @@
 import 'bulma/css/bulma.css';
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App.js";
+import App from "./components/App/App";
 
 ReactDOM.render(
   <App />,
